Squircle, fix nav location line-height, update content, reconsider sidebar
This commit does several things, especially the creation of Squircles
for the icon links.
This commit makes some minor content updates for the about-me section.
This commit reconsiders the use of a side-nav for the skills and instead
utilizes the squircles to create a new section.
This commit fixes the nav-link redirect location points by setting
intro-wrapper height to initial.