Updated Skills Section
This commit includes a new icon-font library called dev-icon, which
links to numerous icons of relevant technologies I know. A few things
are missing however, most importantly Ember.js and InDesign. Consider
using this for now and expanding in the future using icomoon.
Additionally, none of the links these icons are attached to actually
work - they redirect to the skills section. Consider creating a modal
with a brief description of each skill in the near future.